to follow PSR standar the exceptions were moved to their own files
