+++ library/Rain/Tpl/Exception.php
@@ -1,0 +1,26 @@
+<?php
+namespace Rain\Tpl;
+
+/**
+ * Basic Rain\Tpl exception.
+ */
+class Exception extends \Exception{
+		
+	/**
+	 * @var Path of template file with error.
+	 */
+	protected $templateFile = '';
+
+	/**
+	 * Handles path of template file with error.
+	 *
+	 * @param string | null $templateFile
+	 * @return \Rain\Tpl_Exception | string
+	 */
+	public function templateFile($templateFile){
+		if(is_null($file)) return $this->templateFile;
+
+		$this->templateFile = (string) $templateFile;
+		return $this;
+	}
+}
