+++ library/Rain/Tpl/NotFoundException.php
@@ -1,0 +1,10 @@
+<?php
+namespace Rain\Tpl;
+
+require_once 'Exception.php';
+
+/**
+ * Exception thrown when template file does not exists.
+ */
+class NotFoundException extends Exception{
+}
